Add ThickSurf property to surfaces and propagate to CFDMesh
This flag serves the same purpose as the MeshGeom variant.

ThickSurf could also be WatertightSurf -- it indicates whether a surface
or TMesh is a part of a watertight body or not.  Yes means that it can
be used for inside / outside detection for CSG operations.  No means
that inside / outside have no meaning and it should not participate in
CSG classification as the decider (can still be decided upon).

This commit adds the mechanism, but does not go through and tag any
surfaces as ThickSurf=false (true is set as a default).  So, a followup
commit(s) to change appropriate surfaces (wake, actuator disk, structure,
etc) to set this flag is appropriate.

Since these features currently already work, this change is not essential
and will not be done at this time.