Move implementation of ColResizeBrowser::change_cursor into GuiDevice.cpp
This was causing an odd error when we updated to FLTK 1.3.5.  Along with
the error was a complaint about forward declaration of Fl_Window.

Since we normally try to avoid implementation of anything in the *.h, I
thought moving it to the *.cpp might help resolve things - perhaps somehow
connected to the forward declaration complaint.

In general, a bunch of ColResizeBrowser is implemented in the
*.h instead of the *.cpp.  Also, all other cases where we make a
specialized version of an Fl_ class are at the top of GuiDevice.h.  I would
consider moving this to the top of the file and prefixing it with Vsp like
VspSlider, VspButton, Vsp_Group -- or even making it a stand-alone pair
of files as this code is _substantially_ more complex than things like
VspSlider.