Clean up TestAreaProj
This test was derived from other analysis tests -- all were meant as more
of a tutorial than a unit test.  Consequently, they did a lot of things
that were not needed and also were very verbose on output.

This reduces that somewhat -- trying to better blend the test/tutorial
nature.

A potential long-term solution would be to separate test vs tutorial
scripts.