Set up pytest within CTest properly
PyTest will be run 'in' the source directory.  This makes it easy for it
to find any required support files.  However, if any of the tests write
any files, they will pollute the source directory.

In that case, the tests and support files should probably be copied to
the build directory.